+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 Modulsammlung - Changelog
 ===
 
-### Versionsinfos erst ab Release 1.0 :-))
+#### 18.12.2015 Version 2.7-beta
+
+* Update: Abstand oder Trennlinie mit/ohne Grafik
+  * CSS ID vergeben
+  * Umstellug: project_vars
+  * Erklärung aktualisiert
+* Update: Alle Bilder mit aus dem Medienpool mit Copyright anzeigen
+  * CSS ID vergeben
+* Update: Google Maps Modul / Routenplaner
+  * Name geändert: Google Maps / Routenplaner
+  * Modulinfo ausgelager
+  * CSS Info ausgelagert
+* Update: Text / Bild / Video / Link Modul
+  * Umstellug: project_vars
+  * Prüfung ob Redactor / MarkItUp Profile schon vorhanden sind
+* Readme aktualisuert
+* Infoseite aktualisiert